Move product card from ui/ to its own top-level directory
The product card files violated the template rule that ui/ components
must not import domain types or call useTranslations. Moved them to
components/product-card/ as a proper feature component directory.
