+++ apps/template/components/product-card/index.ts
@@ -1,0 +1,12 @@
+export { ProductCard, type ProductCardProps, ProductCardSkeleton } from "./product-card";
+export {
+  ProductCard as ProductCardRoot,
+  ProductCardBadge,
+  ProductCardContent,
+  ProductCardImage,
+  ProductCardImageContainer,
+  ProductCardPrice,
+  ProductCardTitle,
+} from "./primitives";
+export { ProductCardQuickAdd } from "./quick-add";
+export { ProductCardSlideshow } from "./slideshow";
