+++ apps/template/components/product-card/primitives.tsx
@@ -3,7 +3,7 @@ import type * as React from "react";
 
 import { Price } from "@/components/product/price";
 import { DiscountBadge } from "@/components/ui/discount-badge";
-import { ProductCardSlideshow } from "@/components/ui/product-card-slideshow";
+import { ProductCardSlideshow } from "./slideshow";
 import { cn } from "@/lib/utils";
 
 interface ProductCardProps extends React.ComponentProps<"article"> {
