+++ apps/template/components/product-card/slideshow.tsx
